// scrape/sofa/types.ts
export interface PlayerEvents {
    events:           Event[];
    hasNextPage:      boolean;
    playedForTeamMap: { [key: string]: number };
    statisticsMap:    { [key: string]: StatisticsMap };
    incidentsMap:     { [key: string]: IncidentsMap };
    onBenchMap:       { [key: string]: boolean };
}

export interface Event {
    tournament:                Tournament;
    roundInfo?:                RoundInfo;
    customId:                  string;
    status:                    Status;
    winnerCode:                number;
    homeTeam:                  Team;
    awayTeam:                  Team;
    homeScore:                 Score;
    awayScore:                 Score;
    time:                      Time;
    changes:                   Changes;
    hasGlobalHighlights:       boolean;
    hasXg?:                    boolean;
    hasEventPlayerStatistics?: boolean;
    hasEventPlayerHeatMap?:    boolean;
    detailId?:                 number;
    id:                        number;
    startTimestamp:            number;
    slug:                      string;
    finalResultOnly:           boolean;
    awayRedCards?:             number;
    coverage?:                 number;
    homeRedCards?:             number;
    aggregatedWinnerCode?:     number;
    previousLegEventId?:       number;
}

export interface Score {
    current?:    number;
    display?:    number;
    period1?:    number;
    period2?:    number;
    normaltime?: number;
    extra1?:     number;
    extra2?:     number;
    overtime?:   number;
    penalties?:  number;
    aggregated?: number;
}

export interface Team {
    name:       string;
    slug:       string;
    shortName:  string;
    gender:     Gender;
    sport:      Sport;
    userCount:  number;
    nameCode:   string;
    disabled?:  boolean;
    national:   boolean;
    type:       number;
    id:         number;
    subTeams:   any[];
    teamColors: TeamColors;
    ranking?:   number;
}

export enum Gender {
    M = "M",
}

export interface Sport {
    name: SportName;
    slug: SportSlug;
    id:   number;
}

export enum SportName {
    Football = "Football",
}

export enum SportSlug {
    Football = "football",
}

export interface TeamColors {
    primary:   string;
    secondary: string;
    text:      string;
}

export interface Changes {
    changes?:        Change[];
    changeTimestamp: number;
}
// ...
export interface RoundInfo {
    round:         number;
    name?:         RoundInfoName;
    slug?:         string;
    cupRoundType?: number;
}

export enum RoundInfoName {
    Final = "Final",
    MatchFor3RDPlace = "Match for 3rd place",
    Quarterfinal = "Quarterfinal",
    Quarterfinals = "Quarterfinals",
    RoundOf16 = "Round of 16",
    Semifinal = "Semifinal",
    Semifinals = "Semifinals",
    The18 = "1/8",
}

export interface Status {
    code:        number;
    description: Description;
    type:        Type;
}

export enum Description {
    Aet = "AET",
    Ap = "AP",
    Canceled = "Canceled",
    Ended = "Ended",
    NotStarted = "Not started",
    Removed = "Removed",
}

export enum Type {
    Canceled = "canceled",
    Finished = "finished",
    Notstarted = "notstarted",
}

export interface Time {
    injuryTime1?:                 number;
    injuryTime2?:                 number;
    currentPeriodStartTimestamp?: number;
    injuryTime4?:                 number;
    injuryTime3?:                 number;
    current?:                     number;
}

export interface Tournament {
    name:             string;
    slug:             string;
    category:         Category;
    uniqueTournament: UniqueTournament;
    priority:         number;
    id:               number;
}
// ...
export interface IncidentsMap {
    goals?:        number;
    assists?:      number;
    yellowCards?:  number;
    penaltyGoals?: number;
    redCards?:     number;
}

export interface StatisticsMap {
    rating: number;
}